Rename routes: `/` -> `/roulette`, `/diff` -> `/`
