+++ web/src/routes/roulette/login/+page.svelte
@@ -11,7 +11,7 @@
     });
 
     function onLogin() {
-        goto("/");
+        goto("/roulette");
     }
 </script>
 
@@ -19,7 +19,7 @@
     <div id="login-container" class="relative flex h-screen flex-col items-center justify-center p-4">
         <Login {onLogin}></Login>
 
-        <a href="/diff" class="absolute bottom-4 flex items-center rounded-md btn-primary px-4 py-2 text-center">
+        <a href="/" class="absolute bottom-4 flex items-center rounded-md btn-primary px-4 py-2 text-center">
             <span class="me-2 iconify octicon--file-diff-16"></span>Looking for the diff viewer?
         </a>
     </div>
